temporally added netdev2 calls to options

+++ drivers/sx1276/sx1276_netdev2.c
@@ -423,7 +423,7 @@ static int _set(netdev2_t *netdev, netopt_t opt, void *val, size_t len)
             sx1276_set_payload_length(dev, *((uint8_t*) val));
             return sizeof(uint8_t);
         case NETOPT_TX_POWER:
-            sx1276_set_power(dev, (uint8_t) *((uint16_t*) val));
+            sx1276_set_power(dev, *((uint8_t*) val));
             return sizeof(uint16_t);
         case NETOPT_LORA_PREAMBLE_LENGTH:
             sx1276_set_preamble_length(dev, *((uint16_t*) val));
